schema: Split into multiple files
The original schema.go contains now the top level structure while the
table, chain and rule parts have been extracted away.
